+++ src/libAtomVM/iff.c
@@ -103,6 +103,9 @@ void scan_iff(const void *iff_binary, int buf_size, unsigned long *offsets, unsi
         } else if (!memcmp(current_record->name, "Type", 4)) {
             offsets[TYPE] = current_pos;
             sizes[TYPE] = ENDIAN_SWAP_32(current_record->size);
+        } else if (!memcmp(current_record->name, "Recs", 4)) {
+            offsets[RECS] = current_pos;
+            sizes[RECS] = ENDIAN_SWAP_32(current_record->size);
         }
 
         current_pos += iff_align(ENDIAN_SWAP_32(current_record->size) + 8);